+++ app/src/main/java/com/example/monopoly/ui/MainActivity.java
@@ -1,7 +1,8 @@
-package com.example.monopoly;
+package com.example.monopoly.ui;
 
 import android.os.Bundle;
 
+import com.example.monopoly.R;
 import com.google.android.material.snackbar.Snackbar;
 
 import androidx.appcompat.app.AppCompatActivity;
